+++ model_development_event_type.go
@@ -11,7 +11,7 @@ package openapi
 
 // DevelopmentEventType struct for DevelopmentEventType
 type DevelopmentEventType struct {
-	Id        int32  `json:"id,omitempty"`
+	Id        int64  `json:"id,omitempty"`
 	Name      string `json:"name,omitempty"`
 	ShortName string `json:"shortName,omitempty"`
 	Group     string `json:"group,omitempty"`
