+++ model_estimate_coverage.go
@@ -11,6 +11,6 @@ package openapi
 
 // EstimateCoverage struct for EstimateCoverage
 type EstimateCoverage struct {
-	CompanyId int32                   `json:"company_id,omitempty"`
+	CompanyId int64                   `json:"company_id,omitempty"`
 	Brokers   EstimateCoverageBrokers `json:"brokers,omitempty"`
 }
