+++ model_grids_meta.go
@@ -11,8 +11,8 @@ package openapi
 
 // GridsMeta struct for GridsMeta
 type GridsMeta struct {
-	TotalRecords     int32  `json:"total_records,omitempty"`
-	RealTotalRecords int32  `json:"real_total_records,omitempty"`
+	TotalRecords     int64  `json:"total_records,omitempty"`
+	RealTotalRecords int64  `json:"real_total_records,omitempty"`
 	State            string `json:"state,omitempty"`
 	NoResultIfLimit  bool   `json:"noResultIfLimit,omitempty"`
 }
