+++ model_industry_country_data.go
@@ -11,6 +11,6 @@ package openapi
 
 // IndustryCountryData struct for IndustryCountryData
 type IndustryCountryData struct {
-	Id   int32  `json:"id,omitempty"`
+	Id   int64  `json:"id,omitempty"`
 	Name string `json:"name,omitempty"`
 }
