+++ model_new.go
@@ -12,7 +12,7 @@ package openapi
 // New struct for New
 type New struct {
 	CompanyId        string `json:"company_id,omitempty"`
-	CiqCompanyId     int32  `json:"ciq_company_id,omitempty"`
+	CiqCompanyId     int64  `json:"ciq_company_id,omitempty"`
 	CompanySymbol    string `json:"company_symbol,omitempty"`
 	Content          string `json:"content,omitempty"`
 	Description      string `json:"description,omitempty"`
@@ -20,6 +20,6 @@ type New struct {
 	OriginalUrl      string `json:"original_url,omitempty"`
 	ProviderName     string `json:"provider_name,omitempty"`
 	Url              string `json:"url,omitempty"`
-	Date             int32  `json:"date,omitempty"`
+	Date             int64  `json:"date,omitempty"`
 	DateDiffFriendly string `json:"date_diff_friendly,omitempty"`
 }
